api/requirements.txt
asyncpg==0.30.0
fastapi==0.116.1
psycopg2-binary== 2.9.10
pydantic==2.11.9
python-dotenv==1.1.0
uvicorn==0.35.0
wheel==0.45.1
yoyo-migrations==9.0.0